fvm: move fvm back to the top-level
avoids some build issues
