+++ cgo/blockstore.go
