+++ cgo/registry.go
